dashboard/requirements.txt
celery==4.3.0
Django==2.2.7
djangorestframework==3.10.3
psycopg2==2.7.3